refactor: rename analytics.ts to analytics-diagnostics.ts
Completes the analytics module split. analytics-query.ts handles results
querying, analytics-diagnostics.ts handles analysis/diagnostics commands.
Updated import paths and stale .md references.

Closes #138
