+++ .canductor/results.tsv
@@ -50,3 +50,4 @@ guard-ref	2026-03-23T23:13:27.937Z	38	block	typecheck:100,tests:0,code_quality:0
 132	2026-03-24T11:12:05.335Z	99	auto_merge	typecheck:100,tests:100,code_quality:94	merged	Verified 132
 133	2026-03-24T11:16:47.583Z	98	auto_merge	typecheck:100,tests:100,code_quality:93	merged	Verified 133
 137	2026-03-24T12:08:14.459Z	99	auto_merge	typecheck:100,tests:100,code_quality:96	merged	Verified 137
+138	2026-03-24T12:10:53.254Z	99	auto_merge	typecheck:100,tests:100,code_quality:97	pending	Verified 138
