+++ packages/cli/src/commands/analytics-diagnostics.ts
